pyproject.toml
[project]
name = "membrane-env"
version = "0.3.0"
description = "OpenEnv environment: Membrane — privacy-aware multi-agent coordination"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "openenv-core>=0.2.2",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio", "pyyaml>=6.0.0"]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
